Prepare for v1.0
Switched to native package format. The versioning works as follows:
major and minor numbers are for major and minor changes respectively,
and the patch number is for doc updates.

Only major and minor numbers are reflected in the CMakeLists file and
the patch number is added to the changelog file.